+++ spyder/plugins/projects/widgets/projectexplorer.py
